+++ E-3-Commerce-front/src/store/userSlice.jsx
@@ -6,7 +6,7 @@ const baseURL = "http://localhost:3001/";
 export const loginUser = createAsyncThunk("user/login",
   async ({ userData, tokenId }) => {
   try {
-    const response = await axios.post(`${baseURL}login`, userData, { googleToken: tokenId }, {
+    const response = await axios.get(`${baseURL}login`, userData, { googleToken: tokenId }, {
       headers: {
         Accept: 'application/json',
       }
